SecondaryIndex read successful
