+++ DiffernecesOfSummaries/src/main/java/metrics/Metric.java
@@ -1,3 +1,5 @@
+package metrics;
+
 import com.orientechnologies.orient.core.db.ODatabaseSession;
 import results.AvgLinksResult;
 import results.MetricTypes;
